% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_helpers.R
\name{arrayjob_option}
\alias{arrayjob_option}
\title{Arrayjob requirement}
\usage{
arrayjob_option(n = 1L, tc = 100L, stepsize = NULL)
}
\arguments{
\item{n}{An integer which represents number of job-array. If n is 1, arrayjob is not required.}

\item{tc}{An integer. Max mumber of tasks executed simultaneously.}

\item{stepsize}{An integer(option). SGE_TASK_STEPSIZE option of "qsub -t".}
}
\description{
Generates "-t" argument for "qsub".
}
\examples{
arrayjob_option(10L)
}
