% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_helpers.R
\name{as_bash_array}
\alias{as_bash_array}
\title{convert lists, vectors, tibbles into \emph{bash-array}}
\usage{
as_bash_array(..., option = "-a")
}
\arguments{
\item{...}{Lists, vectors, or tibbles. Elements with the same name will be overwritten by the last one.}

\item{option}{An option for declare function of bash. This argument is used for all arguments.}
}
\description{
Vectors or each column of tibbles will be interpretted as a bash-array when your output is read by bash.
}
