% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{build_pipeline}
\alias{build_pipeline}
\title{build a pre-build pipeline}
\usage{
build_pipeline(pipe_name, pipe_dir, n_parallel = 2L, pipe_memory = 3L,
  force = FALSE, make_subdir = TRUE)
}
\arguments{
\item{pipe_name}{A character. Your pipeline name.}

\item{pipe_dir}{A directory path where you intend to build a pipeline.
If it does not exist, it will be made recursively.}

\item{n_parallel}{A integer. The number you run elements of your pipeline simultaneously.}

\item{pipe_memory}{A numeric. Request of memory size par slot for pipeline manager.}

\item{force}{Alogical. Whether to run this function when \emph{pipe_dir} exists and \emph{make_subdir} is TRUE.}

\item{make_subdir}{A logical. Whether to make minimum subdirectories under \emph{pipe_dir}}
}
\description{
build a pre-build pipeline
}
