% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobwatch.R
\name{jobwatch}
\alias{jobwatch}
\title{watch a \emph{qsub} job via \emph{qreport}}
\usage{
jobwatch(x, sys_sleep = 60L, max_repeat = 2L, qsub_args = "",
  qrecall = FALSE, verbose = FALSE, debug = FALSE)
}
\arguments{
\item{x}{A list of your job ID, the path of your qsub file, and the time you execute qsub or time before that.}

\item{sys_sleep}{A numeric. \emph{qreport} interval in seconds.}

\item{max_repeat}{A integer. Total times of trying \emph{qsub} the same file.}

\item{qsub_args}{A character. Additional arguments for \emph{qsub/qrecall}.}

\item{qrecall}{A logical. Whether use \emph{qrecall -file} instead of \emph{qsub} when re-subbing your job.}

\item{verbose}{A logical.}

\item{debug}{A logical.}
}
\value{
Invisible. A list of your final job ID, the path of your qsub file, and the time of final qsub.
}
\description{
watch a \emph{qsub} job via \emph{qreport}
}
