% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{make_qsubfile}
\alias{make_qsubfile}
\title{make a file suitable for \emph{qsub}}
\usage{
make_qsubfile(..., name = "pipeline_child", first_line = binbash(),
  parallel = parallel_option(), arrayjob = arrayjob_option(),
  directory = directory_option(), use_bash_profile = TRUE,
  other_req = character(0))
}
\arguments{
\item{...}{Your codes (default: \emph{bash} codes). Each argument should be a character vector. Multiple arguments and multiple elements will be separated with a line break.}

\item{name}{A character}

\item{first_line}{A character. It is written in the first line.}

\item{parallel}{A character}

\item{arrayjob}{A character}

\item{directory}{A character}

\item{use_bash_profile}{A logical. Whether \emph{source ~/.bash_profile} or not.}

\item{other_req}{A character. Other requirements for \emph{qsub}}
}
\description{
make a file suitable for \emph{qsub}
}
