% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_helpers.R
\name{mem}
\alias{mem}
\title{shortcut for memory requirement}
\usage{
mem(x)
}
\arguments{
\item{x}{A numeric scalar.}
}
\description{
shortcut for memory requirement
}
\examples{
mem(5)
resource("-l", mem(5.3))
}
