% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_helpers.R
\name{parallel_option}
\alias{parallel_option}
\title{Parallel job requirement}
\usage{
parallel_option(env = "def_slot", slot = 1L, memory = 5.3,
  master_memory = NULL, ljob = FALSE, no_rerun = TRUE,
  special_que = NULL, docker_images = NA_character_)
}
\arguments{
\item{env}{A character. Choose one of "def_slot", "mpi", "mpi-fillup", "mpi_4", "mpi_8", "mpi_16", "mpi_24".}

\item{slot}{An integer. The number of slots. If env is either "mpi" or "mpi-fillup", 2-length integer vector representing minimun and maximun nubmer of slots is also accepted.}

\item{memory}{A double. Memory requirement(Gb).}

\item{master_memory}{A double (option). Memory requirement for the master que. If \emph{slot} is 1 or \emph{master_memory} is equal to \emph{memory}, this argument will be ignored.}

\item{ljob}{A logical. Whether need to run more than 2 days. if you require more than 128Gb in total, this option is automatically set as FALSE.}

\item{no_rerun}{A logical. Whether allow to run on rerun ques.}

\item{special_que}{A character (option). Choose one of "cp", "docker", "knl", "gpu", "\emph{groupname}", "exclusive"(equivalent to groupname). If specified, ljob option will be ignored.}

\item{docker_images}{A character (option). Valid only if special_que == "docker".}
}
\description{
Generates parallel-job related arguments for "qsub".
}
\examples{
parallel_option(slot = 4L, memory = 10, master_memory = 5, ljob = TRUE)
}
