% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{pipe}
\alias{pipe}
\alias{\%>\%}
\alias{\%$\%}
\alias{\%T>\%}
\title{Pipe operator}
\description{
See \code{magrittr::\link[magrittr]{\%>\%}}, \code{magrittr::\link[magrittr]{\%$\%}},
and \code{magrittr::\link[magrittr]{\%T>\%}} for details.
}
\keyword{internal}
