% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_commands.R
\name{qreport_tbl}
\alias{qreport_tbl}
\title{get \emph{qreport} results as a tibble}
\usage{
qreport_tbl(ID, begin = NA, user = NA, end = NA)
}
\arguments{
\item{ID}{Job ID or Job Name.}

\item{begin}{A character of \strong{\%Y\%m\%d\%H\%M} format. (optional)}

\item{user}{Your user ID. (optional)}

\item{end}{A character of \strong{\%Y\%m\%d\%H\%M} format. (optional)}
}
\description{
get \emph{qreport} results as a tibble
}
