% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_helpers.R
\name{resource}
\alias{resource}
\title{resource requirement}
\usage{
resource(...)
}
\arguments{
\item{...}{Requirements. Each argument should be a character vector. Multiple arguments and multiple elements will be separated with a space.}
}
\description{
resource requirement
}
\examples{
resource("-l", c("def_slot", "1"))
}
