% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{short_operators}
\alias{short_operators}
\alias{sym}
\alias{syms}
\alias{vars}
\title{Short operators}
\description{
See \code{rlang::\link[rlang]{sym}}, \code{rlang::\link[rlang]{syms}}, and \code{dplyr::\link[dplyr]{vars}} for details.
}
\keyword{internal}
