% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{write_and_qrecall}
\alias{write_and_qrecall}
\title{write out and \emph{qrecall}}
\usage{
write_and_qrecall(..., path = fs::path_home(), log_path = NULL,
  recursive = FALSE, add_time = TRUE)
}
\arguments{
\item{...}{Paths to recall. Each argument should be a character vector. Multiple arguments and multiple elements will be separated with a line break.}

\item{path}{A character. The path to write a file for \emph{qrecall}.}

\item{log_path}{A character (optional). The path to write a stdout of \emph{qrecall}. Default, home directory.}

\item{recursive}{A logical. Whether make parent directory recursively when it does NOT exist.}

\item{add_time}{A logical. Whether add the time you execute this function to path for unique naming.}
}
\value{
Invisible. A list of Job ID, the path you write your file to, and the time you execute this function.
}
\description{
write out and \emph{qrecall}
}
