% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{write_and_qsub}
\alias{write_and_qsub}
\title{write out and \emph{qsub}}
\usage{
write_and_qsub(x, path, recursive = FALSE, add_time = TRUE,
  qsub_args = "")
}
\arguments{
\item{x}{A character. contents of file.}

\item{path}{A character. The path to write a file.}

\item{recursive}{A logical. Whether make parent directory recursively when it does NOT exist.}

\item{add_time}{A logical. Whether add the time you execute this function to path for unique naming.}

\item{qsub_args}{Additional arguments for \emph{qsub}.}
}
\value{
Invisible. A list of Job ID, the path you write your file to, and the time you execute this function.
}
\description{
write out and \emph{qsub}
}
