% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobwatch.R
\name{qreport_tbl}
\alias{qreport_tbl}
\title{get \emph{qreport} results as a tibble}
\usage{
qreport_tbl(ID, begin, user = NA_character_, timeout = Inf)
}
\arguments{
\item{ID}{Job ID.}

\item{begin}{A character of \strong{\%Y\%m\%d\%H\%M} format.}

\item{user}{Your user ID. (optional)}

\item{timeout}{waiting time for \emph{qreport}, in seconds, or as a difftime object.}
}
\description{
get \emph{qreport} results as a tibble
}
