% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobwatch.R
\name{qreport_xml}
\alias{qreport_xml}
\title{get \emph{qreport} results in xml fromat}
\usage{
qreport_xml(ID, begin, user = NA_character_)
}
\arguments{
\item{ID}{Job ID.}

\item{begin}{A character of \strong{\%Y\%m\%d\%H\%M} format.}

\item{user}{Your user ID. (optional)}
}
\description{
get \emph{qreport} results in xml fromat
}
