% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{qsub}
\alias{qsub}
\title{\emph{qsub} a file}
\usage{
qsub(path, qsub_args = character(), qrecall = FALSE)
}
\arguments{
\item{path}{A character. The path to a \emph{qsub/qrecall} file.}

\item{qsub_args}{A character. Additional arguments for \emph{qsub/qrecall}.}

\item{qrecall}{A logical. Whether use qrecall -file instead of qsub.}
}
\value{
Invisible. A list of Job ID, the path you write your file to, and the time you execute this function.
}
\description{
\emph{qsub} a file
}
